INC-539: Extract analytics util functions into a separate file
